refactor(resolume): wire push-audit via integrations, keep over-cap files out of scope
Quality gate (fixed by #482 to count correctly) hard-fails CHANGED files that
exceed the caps. The original wiring touched state/mod.rs (a pre-existing
1117-line god-file, >1000 cap) and added the regression test to tests.rs (which
has pre-existing >120-line test fns). Both are unrelated tech debt that my edits
would have dragged into the failing diff.

Instead:
- attach the audit writer via ResolumeRegistry::attach_audit_writer (idempotent,
  Arc<OnceLock> sink) called from state/integrations.rs::sync_resolume_hosts
  (already a small file) BEFORE hosts spawn — state/mod.rs is untouched.
- move the #483 regression test to its own self-contained resolume/latency_tests.rs
  — tests.rs is untouched.

Pre-existing debt filed as follow-ups (see issues).
